move hook to hooks dir
